requirements.txt
# ============================================================
# KR Stock - Open Architecture Requirements
# ============================================================

# ============================================================
# Web Framework
# ============================================================
fastapi==0.104.0
uvicorn[standard]==0.24.0
flask==3.0.0
gunicorn==21.2.0

# ============================================================
# Database
# ============================================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
# timescaledb==2.13.0  # PostgreSQL extension, not a Python package

# ============================================================
# Celery & Redis
# ============================================================
celery[redis]==5.3.4
redis==4.6.0  # celery[redis] requires redis <5.0.0

# ============================================================
# Data Processing
# ============================================================
pandas==2.1.3
numpy==1.26.2
pykrx==1.0.45
yfinance==0.2.32
# FinanceDataReader - Requires Python <3.11, skipped for now

# ============================================================
# AI/LLM
# ============================================================
google-generativeai==0.3.2
openai==1.6.1

# ============================================================
# Utilities
# ============================================================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
tqdm==4.66.1
beautifulsoup4==4.12.2
lxml_html_clean==0.1.0
rich==13.7.0  # CLI UI 라이브러리 (run.py)

# ============================================================
# Monitoring
# ============================================================
prometheus-client==0.19.0

# ============================================================
# Testing
# ============================================================
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-timeout==2.2.0  # 테스트 timeout 설정
httpx==0.25.2
locust==2.18.3

# ============================================================
# Code Quality
# ============================================================
ruff==0.1.8
mypy==1.7.1
sqlfluff==3.0.3
types-redis==4.6.0
types-requests==2.31.0

# ============================================================
# Development
# ============================================================
ipython==8.18.1
ipdb==0.13.13
